Move report.py to report_search.py to avoid conflict with external/report
